% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar.R
\docType{data}
\name{edgar}
\alias{edgar}
\title{Emissions EDGAR}
\format{
A data.frame with links to download EDGAR data
\describe{
\item{data}{datasets}
\item{pol}{pollutants}
\item{sector}{Sector}
\item{links}{description}
\item{url}{URL}
\item{year}{Year}
\item{type}{nc, txt or NA}
\item{note}{Notes}
data(edgar)
}
}
\source{
\url{https://edgar.jrc.ec.europa.eu/}
}
\usage{
data(edgar)
}
\description{
Several datasets
}
\keyword{datasets}
