% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eixport.R
\docType{package}
\name{eixport}
\alias{eixport}
\alias{eixport-package}
\title{eixport: a package for exporting emissions}
\description{
This package provides functions to export emissions to other models
}
\details{
\strong{1) Characteristics}

This package includes functions related to WRF-Chem, Brams and RLINE.
Also, includes the data:
\enumerate{
\item data("emis_opt")
\item data("emisco")
\item data("Lights")
\item data("rawprofile")
}
}
