% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emisco.R
\docType{data}
\name{emisco}
\alias{emisco}
\title{Emissions database from vein package (g/ms)}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 7205 rows and 182 columns.}
\usage{
data(emisco)
}
\description{
Emissions data of CO from vein demo. When runingn the demo in the vein
package, it produces an object called 'net' which includes the emissions
for 168 hours of CO in g/h. This object was exported to shapefiles and open
in QGIS to split the line vector by its vertices with the interface with
GRASS inside QGIS (v.split.vert) and max vert per link = 2.
Then the emissions were divided by the length of the link in meters and by
3600 to obtain emissions in g/(ms).
}
\details{
It includes parameters coordinates of initial and final link calculates in
QGIS
}
\keyword{datasets}
