% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gCO.R
\docType{data}
\name{gCO}
\alias{gCO}
\title{Gridded emissions from VEIN demo}
\format{A sf object of lines with 437 rows and 2 variables:
\describe{
\item{V9}{Emissions of CO (g/h) for 08:00-09:00}
\item{geometry}{geometry}
data(gCO)
}}
\source{
\url{https://github.com/atmoschem/vein}
}
\usage{
data(gCO)
}
\description{
Emissions in g/h for morning rush hour.
}
\keyword{datasets}
