% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raster.R
\name{plot_raster}
\alias{plot_raster}
\title{Plot raster object}
\usage{
plot_raster(
  r,
  log = FALSE,
  min = -3,
  max,
  legend.shrink = 0.98,
  legend.width = 3,
  axe = !llaxis,
  llaxis = F,
  int = 10,
  proj = FALSE,
  col = c("white", colorRampPalette(colors = c("#D1F5B1", "#FFDE24FC", "#C70000"))(39)),
  x_adjust = 0,
  y_adjust = 0,
  zlim = c(cellStats(r, "min"), cellStats(r, "max")),
  hard_zlim = TRUE,
  ...
)
}
\arguments{
\item{r}{raster}

\item{log}{TRUE to plot in log-scale}

\item{min}{log of minimum for plot (default is -3)}

\item{max}{log of maximum for plot}

\item{legend.shrink}{legend height (default is 0.98)}

\item{legend.width}{legend width (default is 3)}

\item{axe}{to plot axis}

\item{llaxis}{to plot custom axis}

\item{int}{argument passed to latitude / longitude functions}

\item{proj}{TRUE to project the raster to latlon}

\item{col}{color}

\item{x_adjust}{to raster shift dx}

\item{y_adjust}{to raster shift dy}

\item{zlim}{zlimits to be passed to plot}

\item{hard_zlim}{bolean, default TRUE, use the maximum color if value is higher than lim\link{2} and lower than lim\link{1}}

\item{...}{arguments to be passing to stats and plot}
}
\description{
functions that modified plot from raster package
}
\examples{
m <- readRDS(paste0(system.file("extdata",package="hackWRF"),"/model.Rds"))


}
