% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawprofile.R
\docType{data}
\name{rawprofile}
\alias{rawprofile}
\title{Raw profile}
\format{A matrix with 1 column and 168 rows
\describe{
data(rawprofile)
}}
\usage{
data(rawprofile)
}
\description{
Raw profile
}
\keyword{datasets}
