% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regrid}
\alias{regrid}
\title{Read EDGAR emissions for exporting to other formats}
\usage{
regrid(edgar, gtarget, ncol, nrow)
}
\arguments{
\item{edgar}{Edgar NetCDF.}

\item{gtarget}{sf grid, which can be produced using wrf_grid.}

\item{ncol}{Number of columns of gtarget.}

\item{nrow}{Number of rows of gtarget.}
}
\description{
Return a raster
}
\examples{
\dontrun{
# Do not run
# IN DEVELOPMENT
}
}
