% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfx_explode.R
\name{sfx_explode}
\alias{sfx_explode}
\title{splits line by vertex}
\usage{
sfx_explode(x)
}
\arguments{
\item{x}{sf LINESTRING.}
}
\value{
spatial lines
}
\description{
\code{\link{sfx_explode}} splits line by vertex
}
\examples{
{
data(emisco)
dim(emisco)
dfco <- sfx_explode(emisco)
dim(dfco)
}
}
