% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cmaq.R
\name{to_cmaq}
\alias{to_cmaq}
\title{Creating text files for runnign CMAQ}
\usage{
to_cmaq(x, ...)
}
\arguments{
\item{x}{file}

\item{...}{ignored}
}
\description{
Create an emission file from wrfinput
}
\note{
in development. Needs description for creating grid vein::make_grid
and other information
}
\examples{
\dontrun{
}
}
