% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_create.R
\name{wrf_create}
\alias{wrf_create}
\title{Create emission files to the WRF-Chem}
\usage{
wrf_create(wrfinput_dir = getwd(), wrfchemi_dir = getwd(),
  domains = 1, frames_per_auxinput5 = 1, auxinput5_interval_m = 60,
  day_offset = 0, io_style_emissions = 2, kemit = 1,
  variaveis = "ecbmz_mosaic", n_aero = 15, COMPRESS = NA,
  force_ncdf4 = FALSE, verbose = FALSE)
}
\arguments{
\item{wrfinput_dir}{folder with the wrfinput file(s)}

\item{wrfchemi_dir}{output folder}

\item{domains}{domain or domains to process}

\item{frames_per_auxinput5}{value from wrf &time_control namelist.input,
number of times in a single emission file}

\item{auxinput5_interval_m}{value from wrf &time_control namelist.input,
interval in minutes between different times}

\item{day_offset}{number of days (can be a fraction) to create multiple files}

\item{io_style_emissions}{from wrf &chem namelist.input}

\item{kemit}{from wrf &chem namelist.input, number of levels of the emission file}

\item{variaveis}{emission species, can be used data(emis_opt)}

\item{n_aero}{number of aerosol species}

\item{COMPRESS}{integer between 1 (least compr) and 9 (most compr) or NA for
no compression}

\item{force_ncdf4}{force NetCDF4 format}

\item{verbose}{print file info}
}
\description{
Create an emission file from wrfinput file(s),
}
\note{
using io_style_emissions = 1, the 00z will be generated with day_offset = 0 and
12z with day_offset = 0.5

Windows users need to rename the emission files
from 'wrfchemi_d01_2011-08-01_00_00_00' to 'wrfchemi_d01_2011-08-01_00:00:00'
or change in namelist the defoult filename before run wrf.exe with these files
}
\examples{
\dontrun{
# Do not run

# emissions for a 1 day forecast for domains 1 and 2

dir.create(file.path(tempdir(), "EMISS"))

# emissions on date_hour style
wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
           wrfchemi_dir         = file.path(tempdir(), "EMISS"),
           domains              = 1:2,
           frames_per_auxinput5 = 25,
           auxinput5_interval_m = 60,
           verbose              = TRUE)

# emissions on 00z / 12z style, create the 00z
wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
           wrfchemi_dir         = file.path(tempdir(), "EMISS"),
           domains              = 1:2,
           io_style_emissions   = 1,
           day_offset           = 0,
           verbose              = TRUE,
           )
# emissions on 00z / 12z style, create the 12z
wrf_create(wrfinput_dir         = system.file("extdata", package = "eixport"),
           wrfchemi_dir         = file.path(tempdir(), "EMISS"),
           domains              = 1:2,
           io_style_emissions   = 1,
           day_offset           = 0.5,
           verbose              = TRUE)
}
}
\seealso{
\code{\link{to_wrf}} and \code{\link{emis_opt}}
}
\author{
Daniel Schuch
}
