% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_get.R
\name{wrf_get}
\alias{wrf_get}
\title{Function to read variables of emission files}
\usage{
wrf_get(file = file.choose(), name = NA)
}
\arguments{
\item{file}{name of file interactively (default) or specified}

\item{name}{name of the variable (any variable)}
}
\description{
Write variable
}
\examples{
\dontrun{
# Do not run

# create the folder and emission file
dir.create("EMISS")
wrf_create(wrfinput_dir = system.file("extdata", package = "eixport"),
         wrfchemi_dir = "EMISS")

# get the name of created file
files <- list.files(path = "EMISS",pattern = "wrfchemi",full.names = T)

# open, put some numbers and write
CO <- wrf_get(file = files[1],name = "E_CO")
CO[] = rnorm(length(CO))
wrf_put(file = files[1],name = "E_CO",POL = CO)
}
}
\seealso{
\code{\link{wrf_plot}} and \code{\link{wrf_put}}
}
\author{
Daniel Schuch
}
