% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_grid.R
\name{wrf_grid}
\alias{wrf_grid}
\title{Creates grid from wrf file}
\usage{
wrf_grid(filewrf, matrix = F, epsg = 4326)
}
\arguments{
\item{filewrf}{wrfinput file (also geo.nc?, other?)}

\item{matrix}{if the output is matrix or polygon (sf)}
}
\description{
Return a spatialfeature polygon or matrix
}
\examples{
\dontrun{
# Do not run
wrf <- paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep="")
gwrf  <- wrf_grid(wrf)
plot(gwrf, axes = T)
}
}
