% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_meta.R
\name{wrf_meta}
\alias{wrf_meta}
\title{Returns metadata (attributes) of wrf file in a data.frame}
\usage{
wrf_meta(file = file.choose())
}
\arguments{
\item{file}{Character; name of file interactively (default) or specified}
}
\description{
\code{\link{wrf_meta}} returns the attributes of a wrf
NetCDF file in a data.frame. Therefore, there is no need to
use ncdump -h "wrf_file"
}
\examples{
{
# Do not run
}
}
