% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_summary.R
\name{wrf_summary}
\alias{wrf_summary}
\title{Summary of WRF files}
\usage{
wrf_summary(file, name, fn = "sum")
}
\arguments{
\item{file}{String path to the wrf.}

\item{name}{String for variables in wrf file, internally read
by \code{\link{wrf_get}} and transformed to raster.}

\item{fn}{String of the function, for instance, "sum"}
}
\value{
data.frame
}
\description{
This return returns a summary for each variable and hour.
\code{\link{wrf_summary}} internally reads the wrf file using
\code{\link{wrf_get}}, then applies raster::cellStats.
}
\examples{
\dontrun{
# do not run
file <- "/home/sergio/R/x86_64-pc-linux-gnu-library/4.0/eixport/extdata/wrfinput_d02"
wrf_summary(file = file,
            name = c("XLAT", "XLONG"),
            fn = "mean")
}
}
