% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfx_explode.R
\name{sfx_explode}
\alias{sfx_explode}
\title{splits line by vertex}
\usage{
sfx_explode(x, ...)
}
\arguments{
\item{x}{spatial lines.}

\item{...}{ignored}
}
\value{
spatial lines
}
\description{
\code{\link{sfx_explode}} splits line by vertex
}
\examples{
{
# Do not run
data(emisco)
emisco <- emisco[1:100, "V8"]
dfco <- sfx_explode(emisco)
etm <- to_munich(sdf = dfco)
names(etm)
class(etm)
head(etm$Emissions)
head(etm$Street)
write.table(x = etm$Emissions, file = paste0(tempfile(), "_Emissions.txt"),
row.names = FALSE, sep = " ", quote = FALSE)
write.table(x = etm$Street, file = paste0(tempfile(), "_Street.txt"),
row.names = FALSE, sep = " ", quote = FALSE)
}
}
\author{
Michael Summer github\link{dot}/hypertidy/silicate/issues/102
}
