% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein_to_wrf.R, R/wrf_grid.R
\name{vein_to_wrf}
\alias{vein_to_wrf}
\alias{vein_to_wrf}
\title{Export emission from vein model to WRF}
\usage{
vein_to_wrf(wrffile, matrix = F)

vein_to_wrf(wrffile, matrix = F)
}
\arguments{
\item{matrix}{if the output is matrix or polygon (sf)}

\item{filewrf}{wrfinput file (also geo.nc?, other?)}

\item{write}{if the emission will be write directly in the file}

\item{verbose}{display adicional information}

\item{filewrf}{wrfinput file (also geo.nc?, other?)}
}
\description{
Return a matrix or write in file

Return a spatialfeature polygon or matrix
}
\examples{
\dontrun{
# Do not run
wrf <- paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep="")
co  <- vein_to_wrf(vein_emission,wrf)
plot(co)
}
\dontrun{
# Do not run
wrf <- paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep="")
gwrf  <- wrf_grid(wrf)
plot(gwrf)
}
}
