% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vein_to_wrf.R
\name{vein_to_wrf}
\alias{vein_to_wrf}
\title{Export emission from vein model to WRF}
\usage{
vein_to_wrf(vemiss, file = file.choose(), write = F, verbose = T)
}
\arguments{
\item{vemiss}{emission from vein package}

\item{file}{wrf emission file}

\item{write}{if the emission will be write directly in the file}

\item{verbose}{display adicional information}
}
\description{
Return a matrix or write in file
}
\examples{
\dontrun{
# Do not run
wrf <- paste(system.file("extdata", package = "EmissV"),"/wrfinput_d02",sep="")
co  <- vein_to_wrf(vein_emission,wrf)
plot(co)
}
}
