% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_summary.R
\name{wrf_summary}
\alias{wrf_summary}
\title{Summary of variables inside WRF files}
\usage{
wrf_summary(file, vars)
}
\arguments{
\item{file}{String path to the wrf.}

\item{vars}{String of WRF variables. If missing, all variables.}
}
\value{
data.frame
}
\description{
This return returns a summary for each variable.
}
\examples{
\dontrun{
# do not run
file = paste0(system.file("extdata", package = "eixport"),"/wrfinput_d02")
wrf_summary(file = file)
}
}
